refactor(fischer): code-splits user cache functions
