+++ fischer/src/cache/user/set.ts
@@ -1,0 +1,9 @@
+import { idCache, nameCache } from './instance';
+import { UserInfo } from './types';
+
+export function set(userInfo: UserInfo) {
+  const { id, username } = userInfo;
+
+  idCache.set(id, userInfo);
+  nameCache.set(username, id);
+}
