+++ fischer/src/cache/user/stats.ts
@@ -1,0 +1,5 @@
+import { idCache } from './instance';
+
+export function count() {
+  return idCache.stats.keys;
+}
